requirements.txt
# Core dependencies for Voice Personality Analysis System

# Audio processing
librosa>=0.11.0
soundfile>=0.12.0
audioread>=3.0.0

# Data analysis and visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning and clustering
scikit-learn>=1.3.0
scipy>=1.11.0

# GUI
tkinter  # Usually comes with Python

# API and web requests
requests>=2.31.0
google-generativeai>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Audio transcription
openai-whisper>=20250625

# File handling
ffmpeg-python>=0.2.0

# Optional: For enhanced topic organization
# These are included in requirements_topic_org.txt if needed 